updated files to be in-sync with the snapshot code download for 14.3" ;

+++ controllers/api/comment-routes.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 const { Comment } = require('../../models');
-const { restore } = require('../../models/User');
+// const { restore } = require('../../models/User');  // snapshot doesn't have it
 
 router.get('/', (req, res) => {
     Comment.findAll({})
@@ -15,10 +15,10 @@ router.post('/', (req, res) => {
     // check the session
     if (req.session) {
         Comment.create({
-        comment_text: req.body.comment_text,
-        user_id: req.session.user_id,
-        // user_id: req.body.user_id,
-        post_id: req.body.post_id
+            comment_text: req.body.comment_text,
+            user_id: req.session.user_id,
+            // user_id: req.body.user_id,
+            post_id: req.body.post_id
         })
         .then(dbCommentData => res.json(dbCommentData))
         .catch(err => {
@@ -29,17 +29,24 @@ router.post('/', (req, res) => {
 });
 
 router.delete('/:id', (req, res) => {
-    Comment.destroy(
-        {
+    if (req.session) {
+        Comment.destroy({
             where: {
                 id: req.params.id
             }
         })
-    .then(dbCommentData => res.json(dbCommentData))
-    .catch(err => {
-        console.log(err);
-        res.status(500).json(err);
-    });
+        .then(dbCommentData => {
+            if (!dbCommentData) {
+                res.status(404).json({ message: 'No comment found with this id!'});
+                return;
+            }
+            res.json(dbCommentData);
+        })
+        .catch(err => {
+            console.log(err);
+            res.status(500).json(err);
+        });
+    }
 });
 
 module.exports = router;
